feat: moves initial section structure
