+++ packages/core/app/components/cms/SectionBoundary.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { Component } from 'react'
 import type { ErrorInfo, ReactNode } from 'react'
