+++ packages/core/src/components/cms/GlobalSections.tsx
@@ -6,7 +6,7 @@ import CUSTOM_COMPONENTS from 'src/customizations/src/components'
 import { PageContentType, getPage } from 'src/server/cms'
 
 import Toast from 'src/components/common/Toast'
-import RenderSections from './RenderSections'
+import RenderSections from '../../../app/components/cms/RenderSections'
 
 import { OverriddenDefaultAlert as Alert } from 'src/components/sections/Alert/OverriddenDefaultAlert'
 import Footer from 'src/components/sections/Footer'
