+++ packages/core/src/pages/[slug]/p.tsx
@@ -10,7 +10,7 @@ import {
   type ServerProductQueryQuery,
   type ServerProductQueryQueryVariables,
 } from '@generated/graphql'
-import RenderSections from 'src/components/cms/RenderSections'
+import RenderSections from 'app/components/cms/RenderSections'
 import BannerNewsletter from 'src/components/sections/BannerNewsletter/BannerNewsletter'
 import { OverriddenDefaultBannerText as BannerText } from 'src/components/sections/BannerText/OverriddenDefaultBannerText'
 import { OverriddenDefaultBreadcrumb as Breadcrumb } from 'src/components/sections/Breadcrumb/OverriddenDefaultBreadcrumb'
